+++ settings.js
@@ -19,6 +19,7 @@ export const wormProfitTrackerOverlayGui = new Gui();
 export const magmaCoreProfitTrackerOverlayGui = new Gui();
 export const abandonedQuarryTrackerOverlayGui = new Gui();
 export const archfiendDiceProfitTrackerOverlayGui = new Gui();
+export const treasureFishingTrackerOverlayGui = new Gui();
 export const fishingProfitTrackerOverlayGui = new Gui();
 
 const categories = ["General", "Chat", "Alerts", "Overlays", "Items and storages", "Rendering", "Commands", "Dev"]
@@ -1181,7 +1182,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     title: "Maximum sea creatures HP count",
     description: "Show maximum N sea creatures nearby (to limit overlay size). Sea creatures with lower HP come first.",
     options: [1, 20],
-    value: 5,
+    value: 6,
     subcategory: "Sea creatures HP"
 })
 .addButton({
@@ -1563,6 +1564,61 @@ Example 2: ${AQUA}/feeshSetTrackerDrops TIKI_MASK 5 2025-05-30 23:59:00`);
 
 .addSwitch({
     category: "Overlays",
+    configName: "treasureFishingTrackerOverlay",
+    title: "Treasure fishing tracker",
+    description: `Shows an overlay with the overview of the treasure fishing catches, and different related statistics.\nDo ${AQUA}/feeshResetTreasureFishing${GRAY} to reset.`,
+    subcategory: "Treasure fishing tracker",
+    value: false
+})
+.addButton({
+    category: "Overlays",
+    configName: "getTreasureDyesSetupHelp",
+    title: "Set Treasure Dyes count",
+    description: "Explains in your chat how to setup Treasure Dyes count and last drop date.",
+    subcategory: "Treasure fishing tracker",
+    onClick() {
+        ChatLib.chat(`
+${LIGHT_PURPLE}${BOLD}Treasure Dyes setup
+
+Do ${AQUA}/feeshSetTrackerDrops <ITEM_ID> <COUNT> <LAST_ON_DATE>${RESET} to initialize your drops history:
+  - <ITEM_ID> is a mandatory item ID - DYE_TREASURE.
+  - <COUNT> is a mandatory number of times you've dropped it.
+  - <LAST_ON_DATE> is optional and, if provided, should be in YYYY-MM-DD hh:mm:ss format. Can not be in future!
+
+Example: ${AQUA}/feeshSetTrackerDrops DYE_TREASURE 2 2025-05-30 23:59:00`);
+    }
+})
+.addSwitch({
+    category: "Overlays",
+    configName: "resetTreasureFishingTrackerOnGameClosed",
+    title: "Reset on closing game",
+    description: "Automatically reset the Treasure fishing tracker when you close Minecraft or reload CT modules.",
+    subcategory: "Treasure fishing tracker",
+    value: false
+})
+.addButton({
+    category: "Overlays",
+    configName: "moveTreasureFishingTrackerOverlay",
+    title: "Move Treasure fishing tracker overlay",
+    description: "Allows to move and resize the overlay text.",
+    subcategory: "Treasure fishing tracker",
+    onClick() {
+        moveOverlay(treasureFishingTrackerOverlayGui);
+    }
+})
+.addButton({
+    category: "Overlays",
+    configName: "resetTreasureFishingTrackerOverlay",
+    title: "Reset Treasure fishing tracker",
+    description: `Resets tracking for Treasure fishing tracker. Executes ${AQUA}/feeshResetTreasureFishing`,
+    subcategory: "Treasure fishing tracker",
+    onClick() {
+        ChatLib.command("feeshResetTreasureFishing noconfirm", true);
+    }
+})
+
+.addSwitch({
+    category: "Overlays",
     configName: "fishingProfitTrackerOverlay",
     title: "Fishing profit tracker",
     description: `
@@ -1829,6 +1885,13 @@ ${GRAY}Do ${AQUA}/feeshResetProfitTracker${GRAY} to reset.`,
     configName: "boxJawbusFollowers",
     title: "Box Jawbus Followers",
     description: "Render box around Jawbus Followers nearby.",
+    subcategory: "Boxing"
+})
+.addSwitch({
+    category: "Rendering",
+    configName: "boxCocoons",
+    title: "Box Cocoons",
+    description: "Render box around Cocoons nearby.",
     subcategory: "Boxing"
 })